--- src/Filament/OdkTeam/Resources/TeamXlsformTemplateResource/Pages/CreateTeamXlsformTemplate.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Stats4sd\FilamentOdkLink\Filament\OdkTeam\Resources\TeamXlsformTemplateResource\Pages;
-
-use Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformTemplateResource\Pages\CreateXlsformTemplate;
-use Stats4sd\FilamentOdkLink\Filament\OdkTeam\Resources\TeamCustomXlsformTemplateResource;
-use Stats4sd\FilamentOdkLink\Filament\OdkTeam\Resources\TeamXlsformTemplateResource;
-
-class CreateTeamXlsformTemplate extends CreateXlsformTemplate
-{
-    protected static string $resource = TeamXlsformTemplateResource::class;
-}
