+++ src/Filament/OdkTeam/Resources/TeamXlsformTemplates/Pages/CreateTeamXlsformTemplate.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Stats4sd\FilamentOdkLink\Filament\OdkTeam\Resources\TeamXlsformTemplateResources\Pages;
+
+use Stats4sd\FilamentOdkLink\Filament\OdkAdmin\Resources\XlsformTemplates\Pages\CreateXlsformTemplate;
+use Stats4sd\FilamentOdkLink\Filament\OdkTeam\Resources\TeamXlsformTemplates\TeamXlsformTemplateResource;
+
+class CreateTeamXlsformTemplate extends CreateXlsformTemplate
+{
+    protected static string $resource = TeamXlsformTemplateResource::class;
+}
